Better the compatibility for autotag
Edited the file treesitter.lua so tags can be auto closed. You can add
more features to it. For it to work, you need to edit the file
"treesitter.lua" and go to the last section and edit. Don't be shy to go
to the official github for options etc.